# *: pass keyspan.Spans by pointer not value