--- a/internal/keyspan/defragment.go
+++ b/internal/keyspan/defragment.go
@@ -21,16 +21,16 @@ const bufferReuseMaxCapacity = 10 << 10 // 10 KB
 type DefragmentMethod interface {
 	// ShouldDefragment takes two abutting spans and returns whether the two
 	// spans should be combined into a single, defragmented Span.
-	ShouldDefragment(cmp base.Compare, left, right Span) bool
+	ShouldDefragment(cmp base.Compare, left, right *Span) bool
 }
 
 // The DefragmentMethodFunc type is an adapter to allow the use of ordinary
 // functions as DefragmentMethods. If f is a function with the appropriate
 // signature, DefragmentMethodFunc(f) is a DefragmentMethod that calls f.
-type DefragmentMethodFunc func(cmp base.Compare, left, right Span) bool
+type DefragmentMethodFunc func(cmp base.Compare, left, right *Span) bool
 
 // ShouldDefragment calls f(cmp, left, right).
-func (f DefragmentMethodFunc) ShouldDefragment(cmp base.Compare, left, right Span) bool {
+func (f DefragmentMethodFunc) ShouldDefragment(cmp base.Compare, left, right *Span) bool {
 	return f(cmp, left, right)
 }
 
@@ -40,7 +40,7 @@ func (f DefragmentMethodFunc) ShouldDefragment(cmp base.Compare, left, right Spa
 // This defragmenting method is intended for use in compactions that may see
 // internal range keys fragments that may now be joined, because the state that
 // required their fragmentation has been dropped.
-var DefragmentInternal DefragmentMethod = DefragmentMethodFunc(func(cmp base.Compare, a, b Span) bool {
+var DefragmentInternal DefragmentMethod = DefragmentMethodFunc(func(cmp base.Compare, a, b *Span) bool {
 	if len(a.Keys) != len(b.Keys) {
 		return false
 	}
@@ -116,7 +116,7 @@ const (
 type DefragmentingIter struct {
 	cmp      base.Compare
 	iter     FragmentIterator
-	iterSpan Span
+	iterSpan *Span
 	iterPos  iterPos
 
 	// curr holds the span at the current iterator position. currBuf is a buffer
@@ -171,16 +171,16 @@ func (i *DefragmentingIter) Close() error {
 
 // SeekGE seeks the iterator to the first span with a start key greater than or
 // equal to key and returns it.
-func (i *DefragmentingIter) SeekGE(key []byte) Span {
+func (i *DefragmentingIter) SeekGE(key []byte) *Span {
 	i.iterSpan = i.iter.SeekGE(key)
-	if !i.iterSpan.Valid() {
+	if i.iterSpan == nil {
 		i.iterPos = iterPosCurr
-		return Span{}
+		return nil
 	}
 	// Save the current span and peek backwards.
-	i.saveCurrent(i.iterSpan)
+	i.saveCurrent()
 	i.iterSpan = i.iter.Prev()
-	if i.iterSpan.Valid() && i.cmp(i.curr.Start, i.iterSpan.End) == 0 && i.checkEqual(i.iterSpan, i.curr) {
+	if i.iterSpan != nil && i.cmp(i.curr.Start, i.iterSpan.End) == 0 && i.checkEqual(i.iterSpan, &i.curr) {
 		// A continuation. The span we originally landed on and defragmented
 		// backwards has a true Start key < key. To obey the FragmentIterator
 		// contract, we must not return this defragmented span. Defragment
@@ -201,11 +201,11 @@ func (i *DefragmentingIter) SeekGE(key []byte) Span {
 
 // SeekLT seeks the iterator to the last span with a start key less than
 // key and returns it.
-func (i *DefragmentingIter) SeekLT(key []byte) Span {
+func (i *DefragmentingIter) SeekLT(key []byte) *Span {
 	i.iterSpan = i.iter.SeekLT(key)
-	if !i.iterSpan.Valid() {
+	if i.iterSpan == nil {
 		i.iterPos = iterPosCurr
-		return Span{}
+		return nil
 	}
 	// Defragment forward to find the end of the defragmented span.
 	i.defragmentForward()
@@ -218,27 +218,27 @@ func (i *DefragmentingIter) SeekLT(key []byte) Span {
 }
 
 // First seeks the iterator to the first span and returns it.
-func (i *DefragmentingIter) First() Span {
+func (i *DefragmentingIter) First() *Span {
 	i.iterSpan = i.iter.First()
-	if !i.iterSpan.Valid() {
+	if i.iterSpan == nil {
 		i.iterPos = iterPosCurr
-		return Span{}
+		return nil
 	}
 	return i.defragmentForward()
 }
 
 // Last seeks the iterator to the last span and returns it.
-func (i *DefragmentingIter) Last() Span {
+func (i *DefragmentingIter) Last() *Span {
 	i.iterSpan = i.iter.Last()
-	if !i.iterSpan.Valid() {
+	if i.iterSpan == nil {
 		i.iterPos = iterPosCurr
-		return Span{}
+		return nil
 	}
 	return i.defragmentBackward()
 }
 
 // Next advances to the next span and returns it.
-func (i *DefragmentingIter) Next() Span {
+func (i *DefragmentingIter) Next() *Span {
 	switch i.iterPos {
 	case iterPosPrev:
 		// Switching directions; The iterator is currently positioned over the
@@ -254,7 +254,7 @@ func (i *DefragmentingIter) Next() Span {
 		// Next once to move onto y, defragment forward to land on the first z
 		// position.
 		i.iterSpan = i.iter.Next()
-		if invariants.Enabled && !i.iterSpan.Valid() {
+		if invariants.Enabled && i.iterSpan == nil {
 			panic("pebble: invariant violation: no next span while switching directions")
 		}
 		// We're now positioned on the first span that was defragmented into the
@@ -262,9 +262,9 @@ func (i *DefragmentingIter) Next() Span {
 		// position's constitutent fragments. In the above example, this would
 		// land on the first 'z'.
 		i.defragmentForward()
-		if !i.iterSpan.Valid() {
+		if i.iterSpan == nil {
 			i.iterPos = iterPosCurr
-			return Span{}
+			return nil
 		}
 
 		// Now that we're positioned over the first of the next set of
@@ -273,20 +273,20 @@ func (i *DefragmentingIter) Next() Span {
 	case iterPosCurr:
 		// iterPosCurr is only used when the iter is exhausted or when the iterator
 		// is at an empty span.
-		if invariants.Enabled && i.iterSpan.Valid() && !i.iterSpan.Empty() {
+		if invariants.Enabled && i.iterSpan != nil && !i.iterSpan.Empty() {
 			panic("pebble: invariant violation: iterPosCurr with valid iterSpan")
 		}
 
 		i.iterSpan = i.iter.Next()
-		if !i.iterSpan.Valid() {
-			return Span{}
+		if i.iterSpan == nil {
+			return nil
 		}
 		return i.defragmentForward()
 	case iterPosNext:
 		// Already at the next span.
-		if !i.iterSpan.Valid() {
+		if i.iterSpan == nil {
 			i.iterPos = iterPosCurr
-			return Span{}
+			return nil
 		}
 		return i.defragmentForward()
 	default:
@@ -295,25 +295,25 @@ func (i *DefragmentingIter) Next() Span {
 }
 
 // Prev steps back to the previous span and returns it.
-func (i *DefragmentingIter) Prev() Span {
+func (i *DefragmentingIter) Prev() *Span {
 	switch i.iterPos {
 	case iterPosPrev:
 		// Already at the previous span.
-		if !i.iterSpan.Valid() {
+		if i.iterSpan == nil {
 			i.iterPos = iterPosCurr
-			return Span{}
+			return nil
 		}
 		return i.defragmentBackward()
 	case iterPosCurr:
 		// iterPosCurr is only used when the iter is exhausted or when the iterator
 		// is at an empty span.
-		if invariants.Enabled && i.iterSpan.Valid() && !i.iterSpan.Empty() {
+		if invariants.Enabled && i.iterSpan != nil && !i.iterSpan.Empty() {
 			panic("pebble: invariant violation: iterPosCurr with valid iterSpan")
 		}
 
 		i.iterSpan = i.iter.Prev()
-		if !i.iterSpan.Valid() {
-			return Span{}
+		if i.iterSpan == nil {
+			return nil
 		}
 		return i.defragmentBackward()
 	case iterPosNext:
@@ -330,7 +330,7 @@ func (i *DefragmentingIter) Prev() Span {
 		// Prev once to move onto y, defragment backward to land on the last x
 		// position.
 		i.iterSpan = i.iter.Prev()
-		if !i.iterSpan.Valid() {
+		if invariants.Enabled && i.iterSpan == nil {
 			panic("pebble: invariant violation: no previous span while switching directions")
 		}
 		// We're now positioned on the last span that was defragmented into the
@@ -341,9 +341,9 @@ func (i *DefragmentingIter) Prev() Span {
 
 		// Now that we're positioned over the last of the prev set of
 		// fragments, defragment backward.
-		if !i.iterSpan.Valid() {
+		if i.iterSpan == nil {
 			i.iterPos = iterPosCurr
-			return Span{}
+			return nil
 		}
 		return i.defragmentBackward()
 	default:
@@ -354,14 +354,14 @@ func (i *DefragmentingIter) Prev() Span {
 // checkEqual checks the two spans for logical equivalence. It uses the passed-in
 // DefragmentMethod and ensures both spans are NOT empty; not defragmenting empty
 // spans is an optimization that lets us load fewer sstable blocks.
-func (i *DefragmentingIter) checkEqual(left, right Span) bool {
-	return (!left.Empty() && !right.Empty()) && i.method.ShouldDefragment(i.cmp, i.iterSpan, i.curr)
+func (i *DefragmentingIter) checkEqual(left, right *Span) bool {
+	return (!left.Empty() && !right.Empty()) && i.method.ShouldDefragment(i.cmp, i.iterSpan, &i.curr)
 }
 
 // defragmentForward defragments spans in the forward direction, starting from
-// i.iter's current position. The span at the current position must be Valid(),
+// i.iter's current position. The span at the current position must be non-nil,
 // but may be Empty().
-func (i *DefragmentingIter) defragmentForward() Span {
+func (i *DefragmentingIter) defragmentForward() *Span {
 	if i.iterSpan.Empty() {
 		// An empty span will never be equal to another span; see checkEqual for
 		// why. To avoid loading non-empty range keys further ahead by calling Next,
@@ -369,16 +369,16 @@ func (i *DefragmentingIter) defragmentForward() Span {
 		i.iterPos = iterPosCurr
 		return i.iterSpan
 	}
-	i.saveCurrent(i.iterSpan)
+	i.saveCurrent()
 
 	i.iterPos = iterPosNext
 	i.iterSpan = i.iter.Next()
-	for i.iterSpan.Valid() {
+	for i.iterSpan != nil {
 		if i.cmp(i.curr.End, i.iterSpan.Start) != 0 {
 			// Not a continuation.
 			break
 		}
-		if !i.checkEqual(i.iterSpan, i.curr) {
+		if !i.checkEqual(i.iterSpan, &i.curr) {
 			// Not a continuation.
 			break
 		}
@@ -388,13 +388,13 @@ func (i *DefragmentingIter) defragmentForward() Span {
 		i.iterSpan = i.iter.Next()
 	}
 	i.curr.Keys = i.keysBuf
-	return i.curr
+	return &i.curr
 }
 
 // defragmentBackward defragments spans in the backward direction, starting from
-// i.iter's current position. The span at the current position must be Valid(),
+// i.iter's current position. The span at the current position must be non-nil,
 // but may be Empty().
-func (i *DefragmentingIter) defragmentBackward() Span {
+func (i *DefragmentingIter) defragmentBackward() *Span {
 	if i.iterSpan.Empty() {
 		// An empty span will never be equal to another span; see checkEqual for
 		// why. To avoid loading non-empty range keys further ahead by calling Next,
@@ -402,16 +402,16 @@ func (i *DefragmentingIter) defragmentBackward() Span {
 		i.iterPos = iterPosCurr
 		return i.iterSpan
 	}
-	i.saveCurrent(i.iterSpan)
+	i.saveCurrent()
 
 	i.iterPos = iterPosPrev
 	i.iterSpan = i.iter.Prev()
-	for i.iterSpan.Valid() {
+	for i.iterSpan != nil {
 		if i.cmp(i.curr.Start, i.iterSpan.End) != 0 {
 			// Not a continuation.
 			break
 		}
-		if !i.checkEqual(i.iterSpan, i.curr) {
+		if !i.checkEqual(i.iterSpan, &i.curr) {
 			// Not a continuation.
 			break
 		}
@@ -421,10 +421,10 @@ func (i *DefragmentingIter) defragmentBackward() Span {
 		i.iterSpan = i.iter.Prev()
 	}
 	i.curr.Keys = i.keysBuf
-	return i.curr
+	return &i.curr
 }
 
-func (i *DefragmentingIter) saveCurrent(span Span) {
+func (i *DefragmentingIter) saveCurrent() {
 	i.currBuf = i.currBuf[:0]
 	i.keysBuf = i.keysBuf[:0]
 	i.keyBuf = i.keyBuf[:0]
@@ -434,15 +434,18 @@ func (i *DefragmentingIter) saveCurrent(span Span) {
 	if cap(i.keyBuf) > bufferReuseMaxCapacity {
 		i.keyBuf = nil
 	}
-	i.curr = Span{
-		Start: i.saveBytes(span.Start),
-		End:   i.saveBytes(span.End),
+	if i.iterSpan == nil {
+		return
 	}
-	for j := range span.Keys {
+	i.curr = Span{
+		Start: i.saveBytes(i.iterSpan.Start),
+		End:   i.saveBytes(i.iterSpan.End),
+	}
+	for j := range i.iterSpan.Keys {
 		i.keysBuf = append(i.keysBuf, Key{
-			Trailer: span.Keys[j].Trailer,
-			Suffix:  i.saveBytes(span.Keys[j].Suffix),
-			Value:   i.saveBytes(span.Keys[j].Value),
+			Trailer: i.iterSpan.Keys[j].Trailer,
+			Suffix:  i.saveBytes(i.iterSpan.Keys[j].Suffix),
+			Value:   i.saveBytes(i.iterSpan.Keys[j].Value),
 		})
 	}
 	i.curr.Keys = i.keysBuf