@jbowens 2022-06-09
Looks like I still have an interleaving iter invariant violation to work through.